% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AaronLun.R
\name{AaronLun}
\alias{AaronLun}
\title{Aaron Lun}
\usage{
AaronLun(...)
}
\description{
Aaron Lun
}
\examples{
library(SummarizedExperiment)
library(iSEE)

se <- SummarizedExperiment()

iSEE(se, initial = list(AaronLun()))
}
