% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharlotteSoneson.R
\name{CharlotteSoneson-class}
\alias{CharlotteSoneson-class}
\alias{initialize,CharlotteSoneson-method}
\alias{.refineParameters,CharlotteSoneson-method}
\alias{.panelColor,CharlotteSoneson-method}
\alias{.fullName,CharlotteSoneson-method}
\alias{.defineOutput,CharlotteSoneson-method}
\alias{.renderOutput,CharlotteSoneson-method}
\title{Charlotte Soneson}
\description{
Charlotte Soneson
}
