% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharlotteSoneson.R
\name{CharlotteSoneson}
\alias{CharlotteSoneson}
\title{Charlotte Soneson}
\usage{
CharlotteSoneson(...)
}
\description{
Charlotte Soneson
}
\examples{
library(SummarizedExperiment)
library(iSEE)

se <- SummarizedExperiment()

iSEE(se, initial = list(CharlotteSoneson()))
}
