% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedericoMarini.R
\name{FedericoMarini-class}
\alias{FedericoMarini-class}
\alias{initialize,FedericoMarini-method}
\alias{.refineParameters,FedericoMarini-method}
\alias{.panelColor,FedericoMarini-method}
\alias{.fullName,FedericoMarini-method}
\alias{.defineOutput,FedericoMarini-method}
\alias{.renderOutput,FedericoMarini-method}
\title{Federico Marini}
\description{
Federico Marini
}
