% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FedericoMarini.R
\name{FedericoMarini}
\alias{FedericoMarini}
\title{Federico Marini}
\usage{
FedericoMarini(...)
}
\description{
Federico Marini
}
\examples{
library(SummarizedExperiment)
library(iSEE)

se <- SummarizedExperiment()

iSEE(se, initial = list(FedericoMarini()))
}
