% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEtheTeam-package.R
\docType{package}
\name{iSEEtheTeam-pkg}
\alias{iSEEtheTeam}
\alias{iSEEtheTeam-package}
\alias{iSEEtheTeam-pkg}
\title{iSEEtheTeam: iSEE extension for summarising data points in hexagonal bins}
\description{
\code{iSEEtheTeam} is a package that provides panels summarising data points in hexagonal bins for \code{iSEE}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://isee.github.io/iSEEDemoEuroBioC2024/}
  \item Report bugs at \url{https://github.com/iSEE/iSEEDemoEuroBioC2024/issues/new/choose}
}

}
\author{
Kevin Rue-Albrecht \email{kevin.rue-albrecht@kennedy.ox.ac.uk}

Charlotte Soneson \email{charlottesoneson@gmail.com}

Federico Marini \email{marinif@uni-mainz.de}

Aaron Lun \email{infinite.monkeys.with.keyboards@gmail.com}
}
\keyword{internal}
