% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KevinRue.R
\name{KevinRue-class}
\alias{KevinRue-class}
\alias{initialize,KevinRue-method}
\alias{.refineParameters,KevinRue-method}
\alias{.panelColor,KevinRue-method}
\alias{.fullName,KevinRue-method}
\alias{.defineOutput,KevinRue-method}
\alias{.renderOutput,KevinRue-method}
\alias{.defineInterface,KevinRue-method}
\title{Kevin Rue}
\description{
Kevin Rue
}
