% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KevinRue.R
\name{KevinRue}
\alias{KevinRue}
\title{Kevin Rue}
\usage{
KevinRue(...)
}
\description{
Kevin Rue
}
\examples{
library(SummarizedExperiment)
library(iSEE)

se <- SummarizedExperiment()

iSEE(se, initial = list(
  KevinRue(PanelWidth=3L, PanelHeight=400L),
  FedericoMarini(PanelWidth=3L, PanelHeight=400L),
  CharlotteSoneson(PanelWidth=3L, PanelHeight=400L),
  AaronLun(PanelWidth=3L, PanelHeight=400L)
))
}
